[process/task] Normalize tasks types and processes
When executing tasks, we now always use node-pty.

The difference between a process or a shell task relies in the way the
command and its arguments are interpreted/executed:

- process: ask the system to spawn a specific executable, with args
- shell: defer the spawning to a shell that will evaluate a command line

When spawning a process, the expected paramaters are of the kind:

	spawn(executable, [arg1, arg2, arg3, ...])

While when spawning a shell task, we expect some command like

	$> executable arg1 "arg 2" ... && sleep 1 || echo fail

This commit makes Theia tasks and "terminal processes" (processes
spawned via node-pty) able to use both inputs.


[process] Mark as killed when process errors out

Removes some unhandled errors from the tests because we used to not mark
processes that errored as `killed`.

Also renamed a potential name collision with the global `process` name.


[process] Add support for arguments in shell task

VS Code supports an extra `args` field for shell tasks, and will try its
best to build a command-line as `task.command ...task.args`, it will
also take the care of escaping some arguments. It is also possible to
explicitly define what escaping style you want the task runner to use.

This commit makes a best effort to comply to most use cases.


---

IMPORTANT: Disabled `Terminal Process { shell: true }` on Windows.

Problem is that only in this scenario, a lot of garbage ends up in the
standard output of the processes, and I have no idea why. In practice
this doesn't impact the actual terminals inside Theia.

So the tests are disabled on Windows by default, you can have them run
by setting an environment variable named `THEIA_PROCESS_TEST_OVERRIDE`.

+++ packages/process/src/node/raw-process.ts
@@ -21,6 +21,10 @@ import { Process, ProcessType, ProcessOptions, ForkOptions, ProcessErrorEvent } 
 import { ChildProcess, spawn, fork } from 'child_process';
 import * as stream from 'stream';
 
+// The class was here before, exporting to not break anything.
+export { DevNullStream } from './dev-null-stream';
+import { DevNullStream } from './dev-null-stream';
+
 export const RawProcessOptions = Symbol('RawProcessOptions');
 
 /**
@@ -50,30 +54,32 @@ export interface RawProcessFactory {
     (options: RawProcessOptions | RawForkOptions): RawProcess;
 }
 
-/**
- * A Node stream like `/dev/null`.
- *
- * Writing goes to a black hole, reading returns `EOF`.
- */
-class DevNullStream extends stream.Duplex {
-    // tslint:disable-next-line:no-any
-    _write(chunk: any, encoding: string, callback: (err?: Error) => void): void {
-        callback();
-    }
-
-    _read(size: number): void {
-        // tslint:disable-next-line:no-null-keyword
-        this.push(null);
-    }
-}
-
 @injectable()
 export class RawProcess extends Process {
 
-    readonly input: stream.Writable;
-    readonly output: stream.Readable;
-    readonly errorOutput: stream.Readable;
-    readonly process: ChildProcess;
+    /**
+     * @deprecated use `inputStream` instead.
+     */
+    get input(): stream.Writable { return this.inputStream; }
+
+    /**
+     * @deprecated use `outputStream` instead.
+     */
+    get output(): stream.Readable { return this.outputStream; }
+
+    /**
+     * @deprecated use `errorStream` instead.
+     */
+    get errorOutput(): stream.Readable { return this.errorStream; }
+
+    /**
+     * If the process fails to launch, it will be undefined.
+     */
+    readonly process: ChildProcess | undefined;
+
+    readonly outputStream: stream.Readable;
+    readonly errorStream: stream.Readable;
+    readonly inputStream: stream.Writable;
 
     constructor(
         @inject(RawProcessOptions) options: RawProcessOptions | RawForkOptions,
@@ -118,35 +124,34 @@ export class RawProcess extends Process {
                 );
             });
 
-            this.output = this.process.stdout || new DevNullStream();
-            this.input = this.process.stdin || new DevNullStream();
-            this.errorOutput = this.process.stderr || new DevNullStream();
+            this.outputStream = this.process.stdout || new DevNullStream();
+            this.inputStream = this.process.stdin || new DevNullStream();
+            this.errorStream = this.process.stderr || new DevNullStream();
 
             if (this.process.pid !== undefined) {
-                process.nextTick(() => {
-                    this.emitOnStarted();
-                });
+                process.nextTick(this.emitOnStarted.bind(this));
             }
         } catch (error) {
             /* When an error is thrown, set up some fake streams, so the client
                code doesn't break because these field are undefined.  */
-            this.output = new DevNullStream();
-            this.input = new DevNullStream();
-            this.errorOutput = new DevNullStream();
+            this.outputStream = new DevNullStream();
+            this.inputStream = new DevNullStream();
+            this.errorStream = new DevNullStream();
 
             /* Call the client error handler, but first give them a chance to register it.  */
-            process.nextTick(() => {
-                this.errorEmitter.fire(error);
-            });
+            this.emitOnErrorAsync(error);
         }
     }
 
     get pid() {
+        if (!this.process) {
+            throw new Error('process did not start correctly');
+        }
         return this.process.pid;
     }
 
     kill(signal?: string) {
-        if (this.killed === false) {
+        if (this.process && this.killed === false) {
             this.process.kill(signal);
         }
     }
